2 main changes on this massive commit
- genomes rrna scripts slight refactor and renamed (I think these names
  better reflect what they do)
- refactor the unit tests to use the executables instead of python -m
