+++ .github/workflows/pytest-workflow.yml
@@ -18,7 +18,7 @@ jobs:
         with:
           python-version: "3.9"
       - name: Install dependencies
-        run: pip install .[tests]
+        run: pip install -e .[tests]
       - uses: pavelzw/pytest-action@v2
         with:
           emoji: false
