+++ mgnify_pipelines_toolkit/analysis/genomes/rna/extract_bacterial_rrnas_as_tsv.py
@@ -1,19 +1,18 @@
 #!/usr/bin/env python
+# -*- coding: utf-8 -*-
 
-# This file is part of MGnify genome analysis pipeline.
+# Copyright 2024-2025 EMBL - European Bioinformatics Institute
 #
-# MGnify genome analysis pipeline is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
-
-# MGnify genome analysis pipeline is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# GNU General Public License for more details.
-
-# You should have received a copy of the GNU General Public License
-# along with MGnify genome analysis pipeline. If not, see <https://www.gnu.org/licenses/>.
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
 
 
 import os
@@ -42,7 +41,9 @@ def get_tblout_column_indices(tool):
 
 
 def main():
-    parser = argparse.ArgumentParser(description="Script detects bacretia rRNA")
+    parser = argparse.ArgumentParser(
+        description="Script detects bacterial 5S/SSU and LSU rRNA"
+    )
     parser.add_argument(
         "-i", "--input", dest="input", help="rrna.tblout.deoverlapped", required=True
     )
@@ -103,6 +104,7 @@ def main():
 
     if args.outfile:
         file_out = open(args.outfile, "w")
+
     for rna in rRNAs_merged.keys():
         totalLen = 0
         for interval in rRNAs_merged[rna]:
@@ -117,6 +119,7 @@ def main():
             file_out.write(new_line)
         else:
             print(new_line)
+
     if args.outfile:
         file_out.close()
