+++ pyproject.toml
@@ -40,10 +40,11 @@ packages = ["mgnify_pipelines_toolkit",
             "mgnify_pipelines_toolkit.analysis.shared",
             "mgnify_pipelines_toolkit.analysis.amplicon",
             "mgnify_pipelines_toolkit.analysis.assembly",
-            ]
+            "mgnify_pipelines_toolkit.analysis.genomes"
+]
 
 [project.scripts]
-# analysis.shared
+# analysis.shared #
 get_subunits = "mgnify_pipelines_toolkit.analysis.shared.get_subunits:main"
 get_subunits_coords = "mgnify_pipelines_toolkit.analysis.shared.get_subunits_coords:main"
 mapseq2biom = "mgnify_pipelines_toolkit.analysis.shared.mapseq2biom:main"
@@ -52,10 +53,8 @@ library_strategy_check = "mgnify_pipelines_toolkit.analysis.shared.library_strat
 study_summary_generator = "mgnify_pipelines_toolkit.analysis.shared.study_summary_generator:cli"
 markergene_study_summary = "mgnify_pipelines_toolkit.analysis.shared.markergene_study_summary:main"
 convert_cmscan_to_cmsearch_tblout = "mgnify_pipelines_toolkit.analysis.shared.convert_cmscan_to_cmsearch_tblout:main"
-parse_rRNA_bacteria = "mgnify_pipelines_toolkit.analysis.shared.parse_rRNA_bacteria.main"
-parse_tRNA = "mgnify_pipelines_toolkit.analysis.shared.parse_tRNA.main"
-rRNA2seq = "mgnify_pipelines_toolkit.analysis.shared.rRNA2seq.main"
-# analysis.amplicon
+dwc_summary_generator = "mgnify_pipelines_toolkit.analysis.shared.dwc_summary_generator:main"
+# analysis.amplicon #
 are_there_primers = "mgnify_pipelines_toolkit.analysis.amplicon.are_there_primers:main"
 assess_inflection_point_mcp = "mgnify_pipelines_toolkit.analysis.amplicon.assess_inflection_point_mcp:main"
 assess_mcp_proportions = "mgnify_pipelines_toolkit.analysis.amplicon.assess_mcp_proportions:main"
@@ -67,13 +66,18 @@ rev_comp_se_primers = "mgnify_pipelines_toolkit.analysis.amplicon.rev_comp_se_pr
 standard_primer_matching = "mgnify_pipelines_toolkit.analysis.amplicon.standard_primer_matching:main"
 mapseq_to_asv_table = "mgnify_pipelines_toolkit.analysis.amplicon.mapseq_to_asv_table:main"
 primer_val_classification = "mgnify_pipelines_toolkit.analysis.amplicon.primer_val_classification:main"
-# analysis.assembly
+# analysis.assembly #
 krona_txt_from_cat_classification = "mgnify_pipelines_toolkit.analysis.assembly.krona_txt_from_cat_classification:main"
 add_rhea_chebi_annotation = "mgnify_pipelines_toolkit.analysis.assembly.add_rhea_chebi_annotation:main"
 combined_gene_caller_merge = "mgnify_pipelines_toolkit.analysis.assembly.combined_gene_caller_merge:main"
 generate_gaf = "mgnify_pipelines_toolkit.analysis.assembly.generate_gaf:main"
 summarise_goslims = "mgnify_pipelines_toolkit.analysis.assembly.summarise_goslims:main"
-dwc_summary_generator = "mgnify_pipelines_toolkit.analysis.assembly.dwc_summary_generator:main"
+antismash_gff_builder = "mgnify_pipelines_toolkit.analysis.assembly.antismash_gff_builder:main"
+# genomes #
+genomes_extract_bacterial_rrnas_as_tsv = "mgnify_pipelines_toolkit.analysis.genomes.rna.extract_bacterial_rrnas_as_tsv:main"
+genomes_extract_bacterial_rrnas_as_fasta = "mgnify_pipelines_toolkit.analysis.genomes.rna.extract_bacterial_rrnas_as_fasta:main"
+genomes_extract_trnas = "mgnify_pipelines_toolkit.analysis.genomes.rna.extract_trnas:main"
+
 # utils
 fasta_to_delimited = "mgnify_pipelines_toolkit.utils.fasta_to_delimited:main"
 get_mpt_version = "mgnify_pipelines_toolkit.utils.get_mpt_version:main"
