+++ pytest.ini
@@ -1,4 +1,4 @@
 # pytest.ini
 [pytest]
-testpaths = tests
+testpaths = tests/unit
 required_plugins = pytest-workflow
