+++ tests/unit/analysis/amplicon/find_mcp_inflection_points/test_find_mcp_inflection_points.yml
@@ -2,21 +2,21 @@
   tags:
     - find_mcp_inflection_points
   exit_code: 2
-  command: python -m mgnify_pipelines_toolkit.analysis.amplicon.find_mcp_inflection_points --chromosome X
+  command: find_mcp_inflection_points --chromosome X
   stderr:
     contains:
-      - "find_mcp_inflection_points.py: error: the following arguments are required: -i/--input, -s/--sample, -o/--output"
+      - "find_mcp_inflection_points: error: the following arguments are required: -i/--input, -s/--sample, -o/--output"
 - name: find_mcp_inflection_points test_correct_args
   tags:
     - find_mcp_inflection_points
-  command: python -m mgnify_pipelines_toolkit.analysis.amplicon.find_mcp_inflection_points -i tests/fixtures/misc/test_mcp_cons.tsv -s sample -o ./
+  command: find_mcp_inflection_points -i tests/fixtures/misc/test_mcp_cons.tsv -s sample -o ./
   files:
     - path: "sample_inf_points.tsv"
       md5sum: b4da973b056b9b5ed4ec7ca8d7998be4
 - name: find_mcp_inflection_points test_large_file
   tags:
     - find_mcp_inflection_points
-  command: python -m mgnify_pipelines_toolkit.analysis.amplicon.find_mcp_inflection_points -i tests/fixtures/misc/test_mcp_cons_big.tsv -s sample -o ./
+  command: find_mcp_inflection_points -i tests/fixtures/misc/test_mcp_cons_big.tsv -s sample -o ./
   files:
     - path: "sample_inf_points.tsv"
       md5sum: 80767a078f817aab1b9c67736d14093b
