+++ tests/unit/analysis/amplicon/remove_ambiguous_reads/test_remove_ambiguous_reads.yml
@@ -2,14 +2,14 @@
   tags:
     - remove_ambiguous_reads
   exit_code: 2
-  command: python mgnify_pipelines_toolkit/analysis/amplicon/remove_ambiguous_reads.py --chromosome X
+  command: remove_ambiguous_reads --chromosome X
   stderr:
     contains:
-      - "remove_ambiguous_reads.py: error: the following arguments are required: -f/--fwd, -s/--sample"
+      - "remove_ambiguous_reads: error: the following arguments are required: -f/--fwd, -s/--sample"
 - name: remove_ambiguous_reads test_correct_args
   tags:
     - remove_ambiguous_reads
-  command: python mgnify_pipelines_toolkit/analysis/amplicon/remove_ambiguous_reads.py -f tests/fixtures/sequences/test_1_withNs.fastq.gz -r tests/fixtures/sequences/test_2_withNs.fastq.gz -s sample
+  command: remove_ambiguous_reads -f tests/fixtures/sequences/test_1_withNs.fastq.gz -r tests/fixtures/sequences/test_2_withNs.fastq.gz -s sample
   files:
     - path: "sample_noambig_1.fastq.gz"
       md5sum: 34834a5a828a6feb102330d4d1e2e135
