+++ tests/unit/analysis/assembly/generate_gaf/test_generate_gaf.yml
@@ -2,14 +2,14 @@
   tags:
     - generate_gaf
   exit_code: 2
-  command: python -m mgnify_pipelines_toolkit.analysis.assembly.generate_gaf --chromosome X
+  command: generate_gaf --chromosome X
   stderr:
     contains:
-      - "generate_gaf.py: error: the following arguments are required: -i/--ips_input, -o/--output"
+      - "generate_gaf: error: the following arguments are required: -i/--ips_input, -o/--output"
 - name: generate_gaf test_correct_args
   tags:
     - generate_gaf
-  command: python -m mgnify_pipelines_toolkit.analysis.assembly.generate_gaf -i tests/fixtures/goslim/ips_out.tsv -o ERRTESTING
+  command: generate_gaf -i tests/fixtures/goslim/ips_out.tsv -o ERRTESTING
   files:
     - path: "ERRTESTING_ips_annotations.gaf"
       contains:
