+++ tests/unit/analysis/genomes/rna/test_genomes_extract_bacterial_rrnas_as_fasta.yml
@@ -1,0 +1,13 @@
+- name: genomes_extract_bacterial_rrnas_as_fasta
+  tags:
+    - genomes
+    - genomes/rna
+    - genomes/rrna
+  command: genomes_extract_bacterial_rrnas_as_fasta -i tests/fixtures/rna_prediction/genome_rRNAs.fasta -d tests/fixtures/rna_prediction/genome.ncrna.deoverlap.tbl -s cmscan -o genome_rRNAs.fasta
+  files:
+    - path: "genome_rRNAs.fasta"
+      md5sum: b8cdc9a29ffb778bf919543ffdd2294b
+      contains:
+        - ">genome_rRNAs__MGYG000296009_20__SSU_rRNA_bacteria_hit-1__1-338_len=338"
+        - ">genome_rRNAs__MGYG000296009_23__LSU_rRNA_bacteria_hit-1__2-1913_len=1912"
+        - ">genome_rRNAs__MGYG000296009_23__5S_rRNA_hit-2__2115-2224_len=110"
