+++ tests/unit/analysis/shared/converst_cmscan_to_cmsearch_tblout/test_convert_cmscan_to_cmsearch.yml
@@ -1,7 +1,8 @@
 - name: convert cmscan table into cmsearch table
   tags:
+    - analysis/shared
     - cmscan_to_cmsearch
-  command: python -m mgnify_pipelines_toolkit.analysis.shared.convert_cmscan_to_cmsearch_tblout -i tests/fixtures/rna_prediction/test.cmscan_matches.tbl -o test.cmsearch_converted.tbl
+  command: convert_cmscan_to_cmsearch_tblout -i tests/fixtures/rna_prediction/test.cmscan_matches.tbl -o test.cmsearch_converted.tbl
   files:
     - path: "test.cmsearch_converted.tbl"
       md5sum: 3d7488f9388f179ab1e1ba45cfafe400
