+++ tests/unit/analysis/shared/get_subunits_coords/test_get_subunits_coords.yml
@@ -2,14 +2,14 @@
   tags:
     - get_subunits_coords
   exit_code: 2
-  command: python -m mgnify_pipelines_toolkit.analysis.shared.get_subunits_coords --chromosome X
+  command: get_subunits_coords --chromosome X
   stderr:
     contains:
-      - "get_subunits_coords.py: error: the following arguments are required: -i/--input, -l/--lsu, -s/--ssu"
+      - "get_subunits_coords: error: the following arguments are required: -i/--input, -l/--lsu, -s/--ssu"
 - name: get_subunits_coords test_correct_args
   tags:
     - get_subunits_coords
-  command: python -m mgnify_pipelines_toolkit.analysis.shared.get_subunits_coords -i tests/fixtures/misc/matched_seqs_with_coords.txt -s SSU -l LSU
+  command: get_subunits_coords -i tests/fixtures/misc/matched_seqs_with_coords.txt -s SSU -l LSU
   files:
     - path: "SSU_coords"
       md5sum: 401222d263116bb145c1a67e6acc4709
