+++ tests/unit/analysis/shared/study_summary_generator/test_study_summary_generator.yml
@@ -2,14 +2,14 @@
   tags:
     - study_summary_generator
   exit_code: 2
-  command: python -m mgnify_pipelines_toolkit.analysis.shared.study_summary_generator --chromosome X
+  command: study_summary_generator --chromosome X
   stderr:
     contains:
       - "Error: No such option: --chromosome"
 - name: study_summary_generator test_correct_args_summarise_withasvs_noninsdc
   tags:
     - study_summary_generator
-  command: python -m mgnify_pipelines_toolkit.analysis.shared.study_summary_generator summarise -a tests/fixtures/study_summary_inputs/amplicon -r tests/fixtures/study_summary_inputs/amplicon/qc_passed_runs_1.csv -p chunk1 --non_insdc
+  command: study_summary_generator summarise -a tests/fixtures/study_summary_inputs/amplicon -r tests/fixtures/study_summary_inputs/amplicon/qc_passed_runs_1.csv -p chunk1 --non_insdc
   files:
     - path: "chunk1_PR2_study_summary.tsv"
       md5sum: 7956b692bb9c03b10a3387306a8351ef
@@ -22,7 +22,7 @@
 - name: study_summary_generator test_correct_args_summarise_withoutasvs_noninsdc
   tags:
     - study_summary_generator
-  command: python -m mgnify_pipelines_toolkit.analysis.shared.study_summary_generator summarise -a tests/fixtures/study_summary_inputs/amplicon -r tests/fixtures/study_summary_inputs/amplicon/qc_passed_runs_2.csv -p chunk2 --non_insdc
+  command: study_summary_generator summarise -a tests/fixtures/study_summary_inputs/amplicon -r tests/fixtures/study_summary_inputs/amplicon/qc_passed_runs_2.csv -p chunk2 --non_insdc
   files:
     - path: "chunk2_PR2_study_summary.tsv"
       md5sum: a23b6a9b2d350e777e045c659e5d1480
@@ -32,7 +32,7 @@
   tags:
     - study_summary_generator
   exit_code: 1
-  command: python -m mgnify_pipelines_toolkit.analysis.shared.study_summary_generator summarise -a tests/fixtures/study_summary_inputs/amplicon -r tests/fixtures/study_summary_inputs/amplicon/qc_passed_runs_1.csv -p chunk1
+  command: study_summary_generator summarise -a tests/fixtures/study_summary_inputs/amplicon -r tests/fixtures/study_summary_inputs/amplicon/qc_passed_runs_1.csv -p chunk1
   stderr:
     contains:
       - "{'run': 'mock_V3-V4_V9_double_amp'}"
@@ -40,7 +40,7 @@
   tags:
     - study_summary_generator
   exit_code: 1
-  command: python -m mgnify_pipelines_toolkit.analysis.shared.study_summary_generator summarise -a tests/fixtures/study_summary_inputs/amplicon -r tests/fixtures/study_summary_inputs/amplicon/invalid_passed_runs.csv -p chunk1 --non_insdc
+  command: study_summary_generator summarise -a tests/fixtures/study_summary_inputs/amplicon -r tests/fixtures/study_summary_inputs/amplicon/invalid_passed_runs.csv -p chunk1 --non_insdc
   stderr:
     contains:
       - "{'status': 'all_resultsssss'}"
@@ -48,7 +48,7 @@
   tags:
     - study_summary_generator
   exit_code: 1
-  command: python -m mgnify_pipelines_toolkit.analysis.shared.study_summary_generator summarise -a tests/fixtures/study_summary_inputs/amplicon -r tests/fixtures/study_summary_inputs/amplicon/invalid_taxranks_runs.csv -p chunk1 --non_insdc
+  command: study_summary_generator summarise -a tests/fixtures/study_summary_inputs/amplicon -r tests/fixtures/study_summary_inputs/amplicon/invalid_taxranks_runs.csv -p chunk1 --non_insdc
   stderr:
     contains:
       - "{'Superkingdom': 'skkkk__Archaea'}"
@@ -57,7 +57,7 @@
 - name: study_summary_generator test_correct_args_merge
   tags:
     - study_summary_generator
-  command: python -m mgnify_pipelines_toolkit.analysis.shared.study_summary_generator merge -a tests/fixtures/study_summary_inputs/amplicon -p merged
+  command: study_summary_generator merge -a tests/fixtures/study_summary_inputs/amplicon -p merged
   files:
     - path: "merged_SILVA-SSU_study_summary.tsv"
       md5sum: 82962f2a87b0a21e3f2226300bc40667
@@ -70,7 +70,7 @@
 - name: study_summary_generator test_singleton_summarise
   tags:
     - study_summary_generator
-  command: python -m mgnify_pipelines_toolkit.analysis.shared.study_summary_generator summarise -a tests/fixtures/study_summary_inputs/amplicon -r tests/fixtures/study_summary_inputs/amplicon/qc_passed_runs_3.csv -p singleton
+  command: study_summary_generator summarise -a tests/fixtures/study_summary_inputs/amplicon -r tests/fixtures/study_summary_inputs/amplicon/qc_passed_runs_3.csv -p singleton
   files:
     - path: "singleton_PR2_study_summary.tsv"
       md5sum: 89dfddaa6d206627f2d3485988c28fec
@@ -83,7 +83,7 @@
 - name: study_summary_generator test_singleton_merge
   tags:
     - study_summary_generator
-  command: python -m mgnify_pipelines_toolkit.analysis.shared.study_summary_generator merge -a tests/fixtures/study_summary_inputs/amplicon/singleton -p singleton-merged
+  command: study_summary_generator merge -a tests/fixtures/study_summary_inputs/amplicon/singleton -p singleton-merged
   files:
     - path: "singleton-merged_SILVA-SSU_study_summary.tsv"
       md5sum: ba9468fdbde331afe860ac37101bdd4c  # identical copy of single run
@@ -92,7 +92,7 @@
 - name: study_summary_generator test_schema_error_handling
   tags:
     - study_summary_generator
-  command: python -m mgnify_pipelines_toolkit.analysis.shared.study_summary_generator summarise -a tests/fixtures/study_summary_inputs/amplicon -r tests/fixtures/study_summary_inputs/amplicon/qc_passed_runs_4.csv -p empty --non_insdc
+  command: study_summary_generator summarise -a tests/fixtures/study_summary_inputs/amplicon -r tests/fixtures/study_summary_inputs/amplicon/qc_passed_runs_4.csv -p empty --non_insdc
   exit_code: 1
   stderr:
     contains:
